==> 01e_FE_Transformer/03_Convert.ipynb <==
import cudf
import os

e = 3

df_score = cudf.read_parquet(
    './data/t4r/train_pred/pred_' + str(e) + '/score.parquet'
).rename(columns={'value': 't4r_score'})
df_rec = cudf.read_parquet(
    './data/t4r/train_pred/pred_' + str(e) + '/rec.parquet'
).rename(columns={'value': 'cand'})

df = df_rec.merge(
    df_score,
    how='left',
    on=['session', 'order']
)
df.rename(columns={'order': 't4r_rank'})

for igfold in range(5):
    os.system('mkdir -p ../data_folds/fold_' + str(igfold) + '/candidates/train/t4r/')
    df.to_parquet('../data_folds/fold_' + str(igfold) + '/candidates/train/t4r/cand.parquet')

df_score = cudf.read_parquet(
    './data/t4r/sub_pred/pred_' + str(e) + '/score.parquet'
).rename(columns={'value': 't4r_score'})
df_rec = cudf.read_parquet(
    './data/t4r/sub_pred/pred_' + str(e) + '/rec.parquet'
).rename(columns={'value': 'cand'})

df = df_rec.merge(
    df_score,
    how='left',
    on=['session', 'order']
)
df.rename(columns={'order': 't4r_rank'})

for igfold in range(5):
    os.system('mkdir -p ../data_folds/fold_' + str(igfold) + '/candidates/sub/t4r/')
    df.to_parquet('../data_folds/fold_' + str(igfold) + '/candidates/sub/t4r/cand.parquet')
